types 폴더 생성후 공통타입 분리
