+++ src/app/_api/owner_api.ts
@@ -1,43 +1,5 @@
 import instance from './instance';
 
-interface Shop {
-  id: string;
-  name: string;
-  category: string;
-  address1: string;
-  address2: string;
-  description: string;
-  imageUrl: string;
-  originalHourlyPay: number;
-  user: {
-    item: {
-      id: string;
-      email: string;
-      type: 'employer' | 'employee';
-      name?: string;
-      phone?: string;
-      address?: string;
-      bio?: string;
-    };
-    href: string;
-  };
-}
-
-interface ShopResponse {
-  item: Shop;
-  links: any[];
-}
-
-interface ShopRequest {
-  name: string;
-  category: string;
-  address1: string;
-  address2: string;
-  description: string;
-  imageUrl: string;
-  originalHourlyPay: number;
-}
-
 // 가게 등록
 export const registerShop = (data: ShopRequest) => {
   return instance.post<ShopResponse>('/shops', data, {
