+++ src/app/_api/worker_api.ts
@@ -1,37 +1,5 @@
 import instance from './instance';
 
-interface UserInfoResponse {
-  item: {
-    id: string;
-    email: string;
-    type: 'employer' | 'employee';
-    name?: string;
-    phone?: string;
-    address?: string;
-    bio?: string;
-    shop?: {
-      item: {
-        id: string;
-        name: string;
-        category: string;
-        address1: string;
-        address2: string;
-        description: string;
-        imageUrl: string;
-        originalHourlyPay: number;
-      };
-    } | null;
-  };
-  links: any[];
-}
-
-interface UserInfoUpdateRequest {
-  name?: string;
-  phone?: string;
-  address?: string;
-  bio?: string;
-}
-
 // 내 정보 조회
 export const getUserInfo = (userId: string) => {
   return instance.get<UserInfoResponse>(`/users/${userId}`, {
