+++ src/app/_types/worker.d.ts
@@ -1,0 +1,31 @@
+interface UserInfoResponse {
+  item: {
+    id: string;
+    email: string;
+    type: 'employer' | 'employee';
+    name?: string;
+    phone?: string;
+    address?: string;
+    bio?: string;
+    shop?: {
+      item: {
+        id: string;
+        name: string;
+        category: string;
+        address1: string;
+        address2: string;
+        description: string;
+        imageUrl: string;
+        originalHourlyPay: number;
+      };
+    } | null;
+  };
+  links: any[];
+}
+
+interface UserInfoUpdateRequest {
+  name?: string;
+  phone?: string;
+  address?: string;
+  bio?: string;
+}
